% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce.R
\name{coalesce_xy}
\alias{coalesce_xy}
\alias{coalesce_yx}
\title{Coalesce helpers}
\usage{
coalesce_xy(x, y)

coalesce_yx(x, y)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
A vector
}
\description{
These are wrappers around \code{dplyr::coalesce}, designed for convenient use in
the \code{conflict} argument of {powerjoin}'s join functions. \code{coalesce_xy()} is
just like \code{dplyr::coalesce} (except it takes only 2 arguments), \code{coalesce_yx()}
looks first in \code{y} and then in \code{x} if \code{y} is missing.
}
\examples{
coalesce_xy(c(NA, 2, 3), c(11, 12, NA))
coalesce_yx(c(NA, 2, 3), c(11, 12, NA))
}
