% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{full_diagnostic}
\alias{full_diagnostic}
\title{Inform on all potential issues}
\usage{
full_diagnostic
}
\description{
This is the output of \code{check_specs()} with all arguments set to \code{"inform"},
it's useful for a complete join diagnostic.
}
