% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bone_op.R
\name{\%==\%}
\alias{\%==\%}
\title{Extended equality operator}
\usage{
x \%==\% y
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\description{
This is the bone operator, it works like \code{==} but \code{NA \%==\% 1} is \code{FALSE} and
\code{NA \%==\% NA} is \code{TRUE}.
}
