% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_and_methods.R
\name{power_left_join}
\alias{power_left_join}
\alias{power_right_join}
\alias{power_inner_join}
\alias{power_full_join}
\title{Power joins}
\usage{
power_left_join(
  x,
  y = NULL,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  keep = NULL,
  na_matches = c("na", "never"),
  check = check_specs(),
  conflict = NULL,
  fill = NULL
)

power_right_join(
  x,
  y = NULL,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  keep = NULL,
  na_matches = c("na", "never"),
  check = check_specs(),
  conflict = NULL,
  fill = NULL
)

power_inner_join(
  x,
  y = NULL,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  keep = NULL,
  na_matches = c("na", "never"),
  check = check_specs(),
  conflict = NULL,
  fill = NULL
)

power_full_join(
  x,
  y = NULL,
  by = NULL,
  copy = FALSE,
  suffix = c(".x", ".y"),
  keep = NULL,
  na_matches = c("na", "never"),
  check = check_specs(),
  conflict = NULL,
  fill = NULL
)
}
\arguments{
\item{x, y}{A pair of data frames, data frame extensions (e.g. a tibble), or
lazy data frames (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{by}{As in \pkg{dplyr}, but extended so user can supply a formula or a list
of character and formulas. Formulas are used for fuzzy joins, see dedicated
section below.}

\item{copy}{Ignored at the moment because \pkg{powerjoin} doesn't support databases}

\item{suffix}{If there are non-joined duplicate variables in \code{x} and
\code{y}, these suffixes will be added to the output to disambiguate them.
Should be a character vector of length 2.}

\item{keep}{A boolean for compatibility with \pkg{dplyr}, or a value among "left",
"right", "both", "none" or "default". See dedicated section below.}

\item{na_matches}{Should two \code{NA} or two \code{NaN} values match?
\itemize{
\item \code{"na"}, the default, treats two \code{NA} or two \code{NaN} values as equal, like
\code{\%in\%}, \code{\link[=match]{match()}}, and \code{\link[=merge]{merge()}}.
\item \code{"never"} treats two \code{NA} or two \code{NaN} values as different, and will
never match them together or to any other values. This is similar to joins
for database sources and to \code{base::merge(incomparables = NA)}.
}}

\item{check}{A list created with \code{check_specs()}}

\item{conflict}{A function, formula, the special value amongst \code{"patch"},
or a named list of such items. If the LHS of the formula is \code{rw} the rhs will
be applied rowwise. Note that the columns will be subsetted with \code{[} so for
list columns \code{.x} or \code{.y} will refer to length 1 lists and you might sometimes
need \code{.x[[1]]} or \code{.y[[1]]}.}

\item{fill}{Values used to replace missing values originating in unmatched keys,
or a named list of such items.}
}
\value{
A data frame
}
\description{
Power joins
}
\section{\code{keep} argument values}{

\itemize{
\item \code{NULL} (default) : merge keys and name them as the left table's keys, and
keep columns used for fuzzy joins from both tables
\item \code{left} : keep only key columns for left table
\item \code{right}: keep only key columns for right table
\item \code{both} or \code{TRUE}: keep key columns from both tables, adding suffix if relevant
\item \code{none} : drop all key columns from the output
\item \code{FALSE} : merge keys and name them as the left table's keys, maps to \code{none} for fuzzy joins
}
}

\section{fuzzy joins}{


To specify fuzzy matching conditions we use formulas in which the we refer to
the columns from the left side data frame using \code{.x} and the right side data frame
using \code{.y}, for instance \code{by = ~ .x$col1 > .y$col2}.

We can specify several condition and even mix equi condition with fuzzy condition,
for instance \code{by = c(col1 = "col2", ~ .x$col3 > .y$col4)}

To fuzzy match strings we can leverage the functions from the \pkg{stringr}
package since they are vectorized on all main arguments,
for instance to match observations where \code{col1} contains \code{col1} we can attach
\pkg{stringr} and do \code{by =  ~ str_detect(.x$col1, fixed(.y$col2))}.

Another useful function is \code{stringdist} from the \pkg{stringdist} package to match
strings that are close enough, for instance \code{by = ~ stringdist::stringdist(.x$a,.y$a) < 2}

We can also define a new column computed during the fuzzy matching, using the
arrow assignment operator, for instance : \code{by = ~ .x$col1 > (mysum <- .y$col2 + .y$col3)}

When the \code{by} condition evaluates to \code{NA} the observation is dismissed. This makes
\code{by = c(a = "b")} slightly different from \code{by = ~ .x$a == .y$b} when \code{na_matches}
is \code{"na"} (the default). To be able to match \code{NA} with \code{NA} in fuzzy matching condition
we can use the \verb{\%==\%} operator (bone operator), defined in this package.
}

\examples{
# See README for a more verbose version
library(tibble)
male_penguins <- tribble(
  ~name,    ~species,     ~island, ~flipper_length_mm, ~body_mass_g,
  "Giordan",    "Gentoo",    "Biscoe",               222L,        5250L,
  "Lynden",    "Adelie", "Torgersen",               190L,        3900L,
  "Reiner",    "Adelie",     "Dream",               185L,        3650L
)

female_penguins <- tribble(
  ~name,    ~species,  ~island, ~flipper_length_mm, ~body_mass_g,
  "Alonda",    "Gentoo", "Biscoe",               211,        4500L,
  "Ola",    "Adelie",  "Dream",               190,        3600L,
  "Mishayla",    "Gentoo", "Biscoe",               215,        4750L,
)

# apply different checks
power_inner_join(
  male_penguins[c("species", "island")],
  female_penguins[c("species", "island")],
  check = check_specs(implicit_keys = "ignore", duplicate_keys_right = "inform")
)

df1 <- tibble(id = 1:3, value = c(10, NA, 30))
df2 <- tibble(id = 2:4, value = c(22, 32, 42))

# handle conflicted columns when joining
power_left_join(df1, df2, by = "id", conflict = `+`)

# the most frequent use case is to coalesce
power_left_join(df1, df2, by = "id", conflict = coalesce_xy)
power_left_join(df1, df2, by = "id", conflict = coalesce_yx)

# the conflict function is applied colwise by default!
power_left_join(df1, df2, by = "id", conflict = ~ sum(.x, .y, na.rm = TRUE))

# apply conflict function rowwise
power_left_join(df1, df2, by = "id", conflict = rw ~ sum(.x, .y, na.rm = TRUE))

# subset columns without repeating keys
power_inner_join(
  male_penguins \%>\% select_keys_and(name),
  female_penguins \%>\% select_keys_and(female_name = name),
  by = c("species", "island")
)

# semi join
power_inner_join(
  male_penguins,
  female_penguins \%>\% select_keys_and(),
  by = c("species", "island")
)

# agregate without repeating keys
power_left_join(
  male_penguins \%>\% summarize_by_keys(male_weight = mean(body_mass_g)),
  female_penguins \%>\% summarize_by_keys(female_weight = mean(body_mass_g)),
  by = c("species", "island")
)

# pack auxiliary colums without repeating keys
power_left_join(
  male_penguins \%>\% pack_along_keys(name = "m"),
  female_penguins \%>\% pack_along_keys(name = "f"),
  by = c("species", "island")
)

# fuzzy join
power_inner_join(
  male_penguins \%>\% select_keys_and(male_name = name),
  female_penguins \%>\% select_keys_and(female_name = name),
  by = c(~.x$flipper_length_mm < .y$flipper_length_mm, ~.x$body_mass_g > .y$body_mass_g)
)

# fuzzy + equi join
power_inner_join(
  male_penguins \%>\% select_keys_and(male_name = name),
  female_penguins \%>\% select_keys_and(female_name = name),
  by = c("island", ~.x$flipper_length_mm > .y$flipper_length_mm)
)

# define new column without repeating computation
power_inner_join(
  male_penguins \%>\% select_keys_and(male_name = name),
  female_penguins \%>\% select_keys_and(female_name = name),
  by = ~ (mass_ratio <- .y$body_mass_g / .x$body_mass_g) > 1.2
)
power_inner_join(
  male_penguins \%>\% select_keys_and(male_name = name),
  female_penguins \%>\% select_keys_and(female_name = name),
  by = ~ (mass_ratio <- .y$body_mass_g / .x$body_mass_g) > 1.2,
  keep = "none"
)

# fill unmatched values
df1 <- tibble(id = 1:3)
df2 <- tibble(id = 1:2, value2 = c(2, NA), value3 = c(NA, 3))
power_left_join(df1, df2, by = "id", fill = 0)
power_left_join(df1, df2, by = "id", fill = list(value2 = 0))

# join recursively
df1 <- tibble(id = 1, a = "foo")
df2 <- tibble(id = 1, b = "bar")
df3 <- tibble(id = 1, c = "baz")
power_left_join(list(df1, df2, df3), by = "id")
power_left_join(df1, list(df2, df3), by = "id")

}
