% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerjoin-package.R
\docType{package}
\name{powerjoin-package}
\alias{powerjoin}
\alias{powerjoin-package}
\title{powerjoin: Extensions of 'dplyr' and 'fuzzyjoin' Join Functions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

We extend 'dplyr' and 'fuzzyjoin' join functions with features to preprocess the data, apply various data checks, and deal with conflicting columns.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/moodymudskipper/powerjoin}
  \item Report bugs at \url{https://github.com/moodymudskipper/powerjoin/issues}
}

}
\author{
\strong{Maintainer}: Antoine Fabri \email{antoine.fabri@gmail.com}

Other contributors:
\itemize{
  \item Hadley Wickham \email{hadley@rstudio.com} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID}) (aut/cre of dplyr) [contributor]
  \item Romain François (\href{https://orcid.org/0000-0002-2444-4226}{ORCID}) (aut of dplyr) [contributor]
  \item David Robinson \email{admiral.david@gmail.com} (aut of fuzzyjoin) [contributor]
  \item RStudio (cph/fnd dplyr) [copyright holder, funder]
}

}
\keyword{internal}
