% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_specs}
\alias{check_specs}
\title{Build a checklist for power joins}
\usage{
check_specs(
  implicit_keys = c("inform", NA, "warn", "abort"),
  column_conflict = c(NA, "inform", "warn", "abort"),
  duplicate_keys_left = c(NA, "inform", "warn", "abort"),
  duplicate_keys_right = c(NA, "inform", "warn", "abort"),
  unmatched_keys_left = c(NA, "inform", "warn", "abort"),
  unmatched_keys_right = c(NA, "inform", "warn", "abort"),
  missing_key_combination_left = c(NA, "inform", "warn", "abort"),
  missing_key_combination_right = c(NA, "inform", "warn", "abort"),
  inconsistent_factor_levels = c(NA, "inform", "warn", "abort"),
  inconsistent_type = c(NA, "inform", "warn", "abort")
)
}
\arguments{
\item{implicit_keys}{WIP}

\item{column_conflict}{WIP}

\item{duplicate_keys_left}{WIP}

\item{duplicate_keys_right}{WIP}

\item{unmatched_keys_left}{WIP}

\item{unmatched_keys_right}{WIP}

\item{missing_key_combination_left}{WIP}

\item{missing_key_combination_right}{WIP}

\item{inconsistent_factor_levels}{WIP}

\item{inconsistent_type}{WIP}
}
\description{
Build a checklist for power joins
}
