% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\docType{data}
\name{full_diagnostic}
\alias{full_diagnostic}
\title{Inform on all potential issues}
\format{
An object of class \code{powerjoin_check} of length 12.
}
\usage{
full_diagnostic
}
\description{
This is the output of \code{check_specs()} with all arguments set to \code{"inform"},
it's useful for a complete join diagnostic.
}
\keyword{datasets}
